requirements.txt
Flask==3.1.0
groq==0.20.0
gtts==2.5.4
pyaudio==0.2.14
pygame==2.5.2
python-dotenv==1.0.1
speechrecognition==3.14.1
